Provide a glossary for the metadata terms employed within the data manager (#38)
Provides a glossary with explanation and examples for the terms used within the data manager. This has 2 major aims:

1.) Allow DM users an easier time in understanding what kind of metadata should be supplied for each field.
2.) Enable DM collaborators to plan their API with the fields that can be expected for each entity object within the Data manager

Open Questions
A) Some of the fields within the data manager seem to be outdated and are not filled regularly such as:

1.) OrganismId in Sample
2.) Ontologyabbreviation in Ontology

B) Should internal identifiers be shown within a glossary? This is relevant for API usage but could be confusing for GUI users.

C) Should there be filled template examples or a template project as a showcase be provided here for easier understanding